Changed how I flipped the screen
So of course the moment you asked me about how I flipped the numbers I
looked up a way to do it in code.  Found a way to do it using
CGAffineTransformMakeRotation(M_PI) and made the change to see if it
worked.  It does and I am kind of happy I found a coding way to do it.